move bitcoin package to ika2btc
